# backend/auth-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Главный API сервиса",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/addtoplaylist": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Этот эндпоинт позволяет добавить трек в плейлист при наличии токенов, trackID и названия плейлиста",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "playlist"
                ],
                "summary": "Добавление трека в плейлист",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access token (format: 'Bearer {token}') из header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Refresh token из cookies",
                        "name": "refresh_token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID добавляемого трека",
                        "name": "trackID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Название плейлиста, в который добавляется трек",
                        "name": "playlistName",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/deletefromplaylist": {
            "delete": {
                "security": [
                    {
                        "CookieAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Этот эндпоинт позволяет удалить трек из плейлиста при наличии токенов, trackID и названия плейлиста",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "playlist"
                ],
                "summary": "Удаление трека из плейлиста",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access token (format: 'Bearer {token}') из header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Refresh token из cookies",
                        "name": "refresh_token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID удаляемого трека",
                        "name": "trackID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Название плейлиста, из которого удаляется трек",
                        "name": "playlistName",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/deleteplaylist": {
            "delete": {
                "security": [
                    {
                        "CookieAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Этот эндпоинт позволяет удалить плейлист при наличии токенов и названия плейлиста",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "playlist"
                ],
                "summary": "Удаление плейлиста",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access token (format: 'Bearer {token}') из header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Refresh token из cookies",
                        "name": "refresh_token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Название удаляемого плейлиста",
                        "name": "playlistName",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/deleteusertrack": {
            "delete": {
                "description": "Этот эндпоинт позволяет удалить трек при наличии токенов",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "track"
                ],
                "summary": "Удаления трека пользователя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access token (format: 'Bearer {token}') из header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Refresh token из cookies",
                        "name": "refresh_token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID трека",
                        "name": "trackID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Владелец трека - Username",
                        "name": "trackOwner",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Жанр трека",
                        "name": "genre",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The track has been deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/emailchecksend": {
            "get": {
                "description": "Проверка наличия имейла в БД при регистрации и авторизации пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authorization"
                ],
                "summary": "Проверка наличия имейла в БД при регистрации и авторизации пользователя (отправка данных со страницы /signup)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "email пользователя",
                        "name": "email",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Наличие или отсутствие email в БД, json(\"exists\": exists(тип bool))",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный метод запроса для проверки email",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка проверки наличия email, ошибка отправки ответа",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/getUserTracks": {
            "get": {
                "description": "Этот эндпоинт позволяет получить треки пользователя при наличии токенов",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Получение треков пользователя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access token (format: 'Bearer {token}') из header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Refresh token из cookies",
                        "name": "refresh_token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Стартовый номер получения плейлистов",
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Конечный номер получения плейлистов",
                        "name": "end",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Метаданные треков",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.TrackMeta"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "main.TrackMeta": {
            "type": "object",
            "properties": {
                "add_to_db_date": {
                    "description": "Дата добавления в БД - не отправлять",
                    "type": "string"
                },
                "album_name": {
                    "type": "string"
                },
                "artist_name": {
                    "type": "string"
                },
                "description": {
                    "description": "Описание песни - необязательное поле",
                    "type": "string"
                },
                "duration": {
                    "description": "Длительность песни - не отправлять",
                    "type": "integer"
                },
                "genre": {
                    "type": "string"
                },
                "likes": {
                    "description": "Количество лайков - не отправлять",
                    "type": "integer"
                },
                "owner": {
                    "description": "Владелец песни - не отправлять",
                    "type": "string"
                },
                "plays": {
                    "description": "Количество прослушиваний - не отправлять",
                    "type": "integer"
                },
                "release_year": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "track_id": {
                    "type": "integer"
                },
                "track_picture": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
    }
}